+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "13.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.6.1",
     "react-scripts": "5.0.0",
     "sass": "^1.57.1",
     "web-vitals": "2.1.4"
