+++ src/App.js
@@ -1,36 +1,31 @@
-import Directory from "./components/category-item/directory/directory.component";
+import Home from './home/home.component.jsx'
+ import {Routes, Route, Outlet} from 'react-router-dom'
 
 const App = () => {
-  const categories = [
-    {
-      "id": 1,
-      "title": "hats",
-      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
-    },
-    {
-      "id": 2,
-      "title": "jackets",
-      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
-    },
-    {
-      "id": 3,
-      "title": "sneakers",
-      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
-    },
-    {
-      "id": 4,
-      "title": "womens",
-      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
-    },
-    {
-      "id": 5,
-      "title": "mens",
-      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
-    }
-]
+
+  const Navigation = () => {
+    return (
+    <div>
+      <h1> Navigation</h1>
+      <Outlet/>
+    </div>
+    )
+  }
+  const Shop = () => {
+    return (<h1>Shop</h1>)
+  }
+ 
   return (
-      <Directory categories={categories}/>
-  );
+    <Routes>
+      <Route path='/' element={<Navigation/>}>
+        <Route index element={ <Home/>}/>
+        <Route path='shop' element={ <Shop/>}/>
+      </Route>
+    </Routes>
+     )
+
+     
+ 
 };
 
 export default App;
